scm-rules/requirements.txt
pan-scm-sdk==0.3.22
PyYAML==6.0.2
